게임 클리어 조건, 문, 열쇠 오브젝트
게임 클리어 조건 : 나간 마리오 6마리
문, 열쇠 오브젝트 추가, 충돌 시 이벤트 구현
테스트를 위해 현재 1스테이지 문과 열쇠를 생성해놓음

+++ NetWorkTermProject/MarioExodus/MarioExodus/Time.cpp
@@ -40,7 +40,7 @@ void Time::Update(float fps)
 		m_dCurframeTime = ::timeGetTime();
 	}
 	
-	m_fElapsedTime = (m_dCurframeTime - m_dPreframeTime) * m_fTimeScale;
+	m_fElapsedTime = (float)(m_dCurframeTime - m_dPreframeTime) * m_fTimeScale;
 
 	m_fTotalframeRate += m_fElapsedTime;
 	m_iCountframeRate++;
